requirements.txt
asgiref==3.7.2
cloudinary==1.38.0
diff-match-patch==20230430
Django==3.2.23
django-cloudinary-storage==0.3.0
django-filter==23.5
django-import-export==3.3.6
djangorestframework==3.14.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pillow==10.2.0
pytz==2023.3.post1
sqlparse==0.4.4
tablib==3.5.0
xlrd==2.0.1
xlwt==1.3.0
